CS162 Final Project/requirements.txt
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cryptography==38.0.4
firebase-admin==6.0.1
Flask-Login==0.6.2
Flask==2.2.2
geographiclib==2.0
geopy==2.3.0
google-api-core==2.11.0
google-api-python-client==2.70.0
google-auth-httplib2==0.1.0
google-auth==2.15.0
google-cloud-core==2.3.2
google-cloud-firestore==2.7.3
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
grpcio-status==1.51.1
grpcio==1.51.1
httplib2==0.21.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
msgpack==1.0.4
proto-plus==1.22.1
protobuf==4.21.12
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.21
pygeohash==1.2.0
PyJWT==2.6.0
pyparsing==3.0.9
python-dotenv==0.21.0
requests==2.28.1
rsa==4.9
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.13
Werkzeug==2.2.2